+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Increment the:
 
 ## [Unreleased]
 
+* [SDK] Add Entity support to Resource
+  [#3652](https://github.com/open-telemetry/opentelemetry-cpp/issues/3652)
 * [BUG] Install a curl seek callback so an OTLP/HTTP export body can be rewound
   when libcurl restarts an upload, instead of failing with
   `CURLE_SEND_FAIL_REWIND` and dropping the batch